move encoding methods
